TIDIJ-222 Renamed file to vehicle_location_service.dart
