+++ fleetcarpooling/lib/VehicleManagamentService/vehicle_managament_service.dart
@@ -3,7 +3,7 @@ import 'package:core/event/vehicle_event.dart';
 import 'package:core/vehicle.dart';
 import 'package:firebase_database/firebase_database.dart';
 import 'package:fleetcarpooling/Models/vehicle_location_model.dart';
-import 'package:fleetcarpooling/services/vehicle_position_service.dart';
+import 'package:fleetcarpooling/services/vehicle_location_service.dart';
 
 abstract class VehicleRepository {
   Future<void> addVehicle(AddVehicleEvent event);
