+++ fleetcarpooling/lib/pages/map_functionality/map.dart
@@ -4,7 +4,7 @@ import 'package:fleetcarpooling/ReservationService/reservation_service.dart';
 import 'package:fleetcarpooling/pages/map_functionality/udp_manager.dart';
 import 'package:core/ui_elements/custom_toast.dart';
 import 'package:fleetcarpooling/pages/map_functionality/udp_message_handler.dart';
-import 'package:fleetcarpooling/services/vehicle_position_service.dart';
+import 'package:fleetcarpooling/services/vehicle_location_service.dart';
 import 'package:fleetcarpooling/ui_elements/vehicle_controller.dart';
 import 'package:flutter/material.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
